Struct flate2::bufread::GzEncoder

pub struct GzEncoder<R> { /* private fields */ }

A gzip streaming encoder

This structure exposes a BufRead interface that will read uncompressed data from the underlying reader and expose the compressed version as a BufRead interface.

Examples

use std::io::prelude::*;
use std::io;
use flate2::Compression;
use flate2::bufread::GzEncoder;
use std::fs::File;
use std::io::BufReader;

// Opens sample file, compresses the contents and returns a Vector or error
// File wrapped in a BufReader implements BufRead

fn open_hello_world() -> io::Result<Vec<u8>> {
    let f = File::open("examples/hello_world.txt")?;
    let b = BufReader::new(f);
    let mut gz = GzEncoder::new(b, Compression::fast());
    let mut buffer = Vec::new();
    gz.read_to_end(&mut buffer)?;
    Ok(buffer)
}

Implementations

impl<R: BufRead> GzEncoder<R>

pub fn new(r: R, level: Compression) -> GzEncoder<R>

Creates a new encoder which will use the given compression level.

The encoder is not configured specially for the emitted header. For header configuration, see the GzBuilder type.

The data read from the stream r will be compressed and available through the returned reader.

impl<R> GzEncoder<R>

pub fn get_ref(&self) -> &R

Acquires a reference to the underlying reader.

pub fn get_mut(&mut self) -> &mut R

Acquires a mutable reference to the underlying reader.

Note that mutation of the reader may result in surprising results if this encoder is continued to be used.

pub fn into_inner(self) -> R

Returns the underlying stream, consuming this encoder