Struct flate2::bufread::ZlibEncoder

pub struct ZlibEncoder<R> { /* private fields */ }

A ZLIB encoder, or compressor.

This structure consumes a BufRead interface, reading uncompressed data from the underlying reader, and emitting compressed data.

Examples

use std::io::prelude::*;
use flate2::Compression;
use flate2::bufread::ZlibEncoder;
use std::fs::File;
use std::io::BufReader;

// Use a buffered file to compress contents into a Vec<u8>

let f = File::open("examples/hello_world.txt")?;
let b = BufReader::new(f);
let mut z = ZlibEncoder::new(b, Compression::fast());
let mut buffer = Vec::new();
z.read_to_end(&mut buffer)?;

Implementations

impl<R: BufRead> ZlibEncoder<R>

pub fn new(r: R, level: Compression) -> ZlibEncoder<R>

Creates a new encoder which will read uncompressed data from the given stream and emit the compressed stream.

impl<R> ZlibEncoder<R>

pub fn reset(&mut self, r: R) -> R

Resets the state of this encoder entirely, swapping out the input stream for another.

This function will reset the internal state of this encoder and replace the input stream with the one provided, returning the previous input stream. Future data read from this encoder will be the compressed version of r’s data.

pub fn get_ref(&self) -> &R

Acquires a reference to the underlying reader

pub fn get_mut(&mut self) -> &mut R

Acquires a mutable reference to the underlying stream

Note that mutation of the stream may result in surprising results if this encoder is continued to be used.

pub fn into_inner(self) -> R

Consumes this encoder, returning the underlying reader.

pub fn total_in(&self) -> u64

Returns the number of bytes that have been read into this compressor.

Note that not all bytes read from the underlying object may be accounted for, there may still be some active buffering.

pub fn total_out(&self) -> u64

Returns the number of bytes that the compressor has produced.

Note that not all bytes may have been read yet, some may still be buffered.