Enum flate2::Compression

pub enum Compression {
    None,
    Fast,
    Best,
    Default,
}

When compressing data, the compression level can be specified by a value in this enum.

Variants

None

No compression is to be performed, this may actually inflate data slightly when encoding.

Fast

Optimize for the best speed of encoding.

Best

Optimize for the size of data being encoded.

Default

Choose the default compression, a balance between speed and size.