Module flate2::gz

Modules

bufread

read

write

Structs

GzBuilder

A builder structure to create a new gzip Encoder.

GzHeader

A structure representing the header of a gzip stream.

GzHeaderParser

Enums

GzHeaderState

Constants

MAX_HEADER_BUF 🔒

Statics

FCOMMENT

FEXTRA

FHCRC

FNAME

FRESERVED

Functions

bad_header 🔒

corrupt 🔒

parse_le_u16 🔒

read_into 🔒

read_to_nul 🔒