flate2

Module read

Types which operate over Read streams, both encoders and decoders for various formats.

Note that the read decoder types may read past the end of the compressed data while decoding. If the caller requires subsequent reads to start immediately following the compressed data wrap the Read type in a BufReader and use the BufReader with the equivalent decoder from the bufread module and also for the subsequent reads.

Structs

DeflateDecoder

A DEFLATE decoder, or decompressor.

DeflateEncoder

A DEFLATE encoder, or compressor.

GzDecoder

A decoder for a single member of a gzip file.

GzEncoder

A gzip streaming encoder

MultiGzDecoder

A gzip streaming decoder that decodes a gzip file that may have multiple members.

ZlibDecoder

A ZLIB decoder, or decompressor.

ZlibEncoder

A ZLIB encoder, or compressor.