Struct flate2::read::DeflateEncoder

pub struct DeflateEncoder<R> { /* fields omitted */ }

A DEFLATE encoder, or compressor.

This structure implements a Read interface and will read uncompressed data from an underlying stream and emit a stream of compressed data.

Examples

use std::io::prelude::*;
use std::io;
use flate2::Compression;
use flate2::read::DeflateEncoder;

// Return a vector containing the Deflate compressed version of hello world
fn deflateencoder_read_hello_world() -> io::Result<Vec<u8>> {
   let mut ret_vec = [0;100];
   let c = b"hello world";
   let mut deflater = DeflateEncoder::new(&c[..], Compression::fast());
   let count = deflater.read(&mut ret_vec)?;
   Ok(ret_vec[0..count].to_vec())
}

Implementations

impl<R: Read> DeflateEncoder<R>

pub fn new(r: R, level: Compression) -> DeflateEncoder<R>

Creates a new encoder which will read uncompressed data from the given stream and emit the compressed stream.

impl<R> DeflateEncoder<R>

pub fn reset(&mut self, r: R) -> R

Resets the state of this encoder entirely, swapping out the input stream for another.

This function will reset the internal state of this encoder and replace the input stream with the one provided, returning the previous input stream. Future data read from this encoder will be the compressed version of r’s data.

Note that there may be currently buffered data when this function is called, and in that case the buffered data is discarded.

pub fn get_ref(&self) -> &R

Acquires a reference to the underlying reader

pub fn get_mut(&mut self) -> &mut R

Acquires a mutable reference to the underlying stream

Note that mutation of the stream may result in surprising results if this encoder is continued to be used.

pub fn into_inner(self) -> R

Consumes this encoder, returning the underlying reader.

Note that there may be buffered bytes which are not re-acquired as part of this transition. It’s recommended to only call this function after EOF has been reached.

pub fn total_in(&self) -> u64

Returns the number of bytes that have been read into this compressor.

Note that not all bytes read from the underlying object may be accounted for, there may still be some active buffering.

pub fn total_out(&self) -> u64

Returns the number of bytes that the compressor has produced.

Note that not all bytes may have been read yet, some may still be buffered.