Struct flate2::read::GzEncoder

pub struct GzEncoder<R> { /* private fields */ }

A gzip streaming encoder

This structure exposes a Read interface that will read uncompressed data from the underlying reader and expose the compressed version as a Read interface.

Examples

use std::io::prelude::*;
use std::io;
use flate2::Compression;
use flate2::read::GzEncoder;

// Return a vector containing the GZ compressed version of hello world

fn gzencode_hello_world() -> io::Result<Vec<u8>> {
    let mut ret_vec = [0;100];
    let bytestring = b"hello world";
    let mut gz = GzEncoder::new(&bytestring[..], Compression::fast());
    let count = gz.read(&mut ret_vec)?;
    Ok(ret_vec[0..count].to_vec())
}

Implementations

impl<R: Read> GzEncoder<R>

pub fn new(r: R, level: Compression) -> GzEncoder<R>

Creates a new encoder which will use the given compression level.

The encoder is not configured specially for the emitted header. For header configuration, see the GzBuilder type.

The data read from the stream r will be compressed and available through the returned reader.

impl<R> GzEncoder<R>

pub fn get_ref(&self) -> &R

Acquires a reference to the underlying reader.

pub fn get_mut(&mut self) -> &mut R

Acquires a mutable reference to the underlying reader.

Note that mutation of the reader may result in surprising results if this encoder is continued to be used.

pub fn into_inner(self) -> R

Returns the underlying stream, consuming this encoder