Struct flate2::read::MultiGzDecoder

pub struct MultiGzDecoder<R> { /* private fields */ }

A gzip streaming decoder that decodes all members of a multistream

A gzip member consists of a header, compressed data and a trailer. The gzip specification, however, allows multiple gzip members to be joined in a single stream. MultiGzDecoder will decode all consecutive members while GzDecoder will only decompress the first gzip member. The multistream format is commonly used in bioinformatics, for example when using the BGZF compressed data.

This structure exposes a Read interface that will consume all gzip members from the underlying reader and emit uncompressed data.

Examples

use std::io::prelude::*;
use std::io;
use flate2::read::MultiGzDecoder;

// Uncompresses a Gz Encoded vector of bytes and returns a string or error
// Here &[u8] implements Read

fn decode_reader(bytes: Vec<u8>) -> io::Result<String> {
   let mut gz = MultiGzDecoder::new(&bytes[..]);
   let mut s = String::new();
   gz.read_to_string(&mut s)?;
   Ok(s)
}

Implementations

impl<R: Read> MultiGzDecoder<R>

pub fn new(r: R) -> MultiGzDecoder<R>

Creates a new decoder from the given reader, immediately parsing the (first) gzip header. If the gzip stream contains multiple members all will be decoded.

impl<R> MultiGzDecoder<R>

pub fn header(&self) -> Option<&GzHeader>

Returns the current header associated with this stream, if it’s valid.

pub fn get_ref(&self) -> &R

Acquires a reference to the underlying reader.

pub fn get_mut(&mut self) -> &mut R

Acquires a mutable reference to the underlying stream.

Note that mutation of the stream may result in surprising results if this decoder is continued to be used.

pub fn into_inner(self) -> R

Consumes this decoder, returning the underlying reader.