Struct flate2::CrcWriter

pub struct CrcWriter<W> { /* fields omitted */ }

A wrapper around a Write that calculates the CRC.

Methods

impl<W> CrcWriter<W>

pub fn crc(&self) -> &Crc

Get the Crc for this CrcWriter.

pub fn into_inner(self) -> W

Get the writer that is wrapped by this CrcWriter.

Important traits for &'a mut W
impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn get_ref(&self) -> &W

Get the writer that is wrapped by this CrcWriter by reference.

Important traits for &'a mut W
impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn get_mut(&mut self) -> &mut W

Get a mutable reference to the writer that is wrapped by this CrcWriter.

pub fn reset(&mut self)

Reset the Crc in this CrcWriter.

impl<W: Write> CrcWriter<W>

Important traits for CrcWriter<W>
impl<W: Write> Write for CrcWriter<W>

pub fn new(w: W) -> CrcWriter<W>

Create a new CrcWriter.