Trait flate2::FlateReadExt

pub trait FlateReadExt: Read + Sized {
    fn gz_encode(self, lvl: Compression) -> GzEncoder<Self> { ... }
    fn gz_decode(self) -> Result<GzDecoder<Self>> { ... }
    fn zlib_encode(self, lvl: Compression) -> ZlibEncoder<Self> { ... }
    fn zlib_decode(self) -> ZlibDecoder<Self> { ... }
    fn deflate_encode(self, lvl: Compression) -> DeflateEncoder<Self> { ... }
    fn deflate_decode(self) -> DeflateDecoder<Self> { ... }
}

A helper trait to create encoder/decoders with method syntax.

Provided Methods

fn gz_encode(self, lvl: Compression) -> GzEncoder<Self>

Consume this reader to create a compression stream at the specified compression level.

fn gz_decode(self) -> Result<GzDecoder<Self>>

Consume this reader to create a decompression stream of this stream.

fn zlib_encode(self, lvl: Compression) -> ZlibEncoder<Self>

Consume this reader to create a compression stream at the specified compression level.

fn zlib_decode(self) -> ZlibDecoder<Self>

Consume this reader to create a decompression stream of this stream.

fn deflate_encode(self, lvl: Compression) -> DeflateEncoder<Self>

Consume this reader to create a compression stream at the specified compression level.

fn deflate_decode(self) -> DeflateDecoder<Self>

Consume this reader to create a decompression stream of this stream.