Module flate2::write

Types which operate over Write streams, both encoders and decoders for various formats.

Structs

DeflateDecoder

A DEFLATE decoder, or decompressor.

DeflateEncoder

A DEFLATE encoder, or compressor.

GzDecoder

A decoder for a single member of a gzip file.

GzEncoder

A gzip streaming encoder

MultiGzDecoder

A gzip streaming decoder that decodes a gzip file with multiple members.

ZlibDecoder

A ZLIB decoder, or decompressor.

ZlibEncoder

A ZLIB encoder, or compressor.