Struct flate2::write::DeflateDecoder

pub struct DeflateDecoder<W: Write> { /* fields omitted */ }

A DEFLATE decoder, or decompressor.

This structure implements a Write and will emit a stream of decompressed data when fed a stream of compressed data.

Methods

impl<W: Write> DecoderWriter<W>

fn new(w: W) -> DecoderWriter<W>

Creates a new decoder which will write uncompressed data to the stream.

When this encoder is dropped or unwrapped the final pieces of data will be flushed.

fn reset(&mut self, w: W) -> Result<W>

Resets the state of this decoder entirely, swapping out the output stream for another.

This function will finish encoding the current stream into the current output stream before swapping out the two output streams. If the stream cannot be finished an error is returned.

This will then reset the internal state of this decoder and replace the output stream with the one provided, returning the previous output stream. Future data written to this decoder will be decompressed into the output stream w.

fn finish(self) -> Result<W>

Consumes this encoder, flushing the output stream.

This will flush the underlying data stream and then return the contained writer if the flush succeeded.

fn total_in(&self) -> u64

Returns the number of bytes that the decompressor has consumed for decompression.

Note that this will likely be smaller than the number of bytes successfully written to this stream due to internal buffering.

fn total_out(&self) -> u64

Returns the number of bytes that the decompressor has written to its output stream.