Struct flate2::write::DeflateEncoder

pub struct DeflateEncoder<W: Write> { /* fields omitted */ }

A DEFLATE encoder, or compressor.

This structure implements a Write interface and takes a stream of uncompressed data, writing the compressed data to the wrapped writer.

Examples

use std::io::prelude::*;
use flate2::Compression;
use flate2::write::DeflateEncoder;

// Vec<u8> implements Write to print the compressed bytes of sample string

let mut e = DeflateEncoder::new(Vec::new(), Compression::default());
e.write_all(b"Hello World").unwrap();
println!("{:?}", e.finish().unwrap());

Implementations

impl<W: Write> DeflateEncoder<W>

pub fn new(w: W, level: Compression) -> DeflateEncoder<W>

Creates a new encoder which will write compressed data to the stream given at the given compression level.

When this encoder is dropped or unwrapped the final pieces of data will be flushed.

pub fn get_ref(&self) -> &W

Acquires a reference to the underlying writer.

pub fn get_mut(&mut self) -> &mut W

Acquires a mutable reference to the underlying writer.

Note that mutating the output/input state of the stream may corrupt this object, so care must be taken when using this method.

pub fn reset(&mut self, w: W) -> Result<W>

Resets the state of this encoder entirely, swapping out the output stream for another.

This function will finish encoding the current stream into the current output stream before swapping out the two output streams. If the stream cannot be finished an error is returned.

After the current stream has been finished, this will reset the internal state of this encoder and replace the output stream with the one provided, returning the previous output stream. Future data written to this encoder will be the compressed into the stream w provided.

Errors

This function will perform I/O to complete this stream, and any I/O errors which occur will be returned from this function.

pub fn try_finish(&mut self) -> Result<()>

Attempt to finish this output stream, writing out final chunks of data.

Note that this function can only be used once data has finished being written to the output stream. After this function is called then further calls to write may result in a panic.

Panics

Attempts to write data to this stream may result in a panic after this function is called.

Errors

This function will perform I/O to complete this stream, and any I/O errors which occur will be returned from this function.

pub fn finish(self) -> Result<W>

Consumes this encoder, flushing the output stream.

This will flush the underlying data stream, close off the compressed stream and, if successful, return the contained writer.

Note that this function may not be suitable to call in a situation where the underlying stream is an asynchronous I/O stream. To finish a stream the try_finish (or shutdown) method should be used instead. To re-acquire ownership of a stream it is safe to call this method after try_finish or shutdown has returned Ok.

Errors

This function will perform I/O to complete this stream, and any I/O errors which occur will be returned from this function.

pub fn flush_finish(self) -> Result<W>

Consumes this encoder, flushing the output stream.

This will flush the underlying data stream and then return the contained writer if the flush succeeded. The compressed stream will not closed but only flushed. This means that obtained byte array can by extended by another deflated stream. To close the stream add the two bytes 0x3 and 0x0.

Errors

This function will perform I/O to complete this stream, and any I/O errors which occur will be returned from this function.

pub fn total_in(&self) -> u64

Returns the number of bytes that have been written to this compresor.

Note that not all bytes written to this object may be accounted for, there may still be some active buffering.

pub fn total_out(&self) -> u64

Returns the number of bytes that the compressor has produced.

Note that not all bytes may have been written yet, some may still be buffered.