Struct flate2::bufread::MultiGzDecoder

pub struct MultiGzDecoder<R>(/* private fields */);

A gzip streaming decoder that decodes a gzip file that may have multiple members.

This structure implements a Read interface. When read from, it reads compressed data from the underlying BufRead and provides the uncompressed data.

A gzip file consists of a series of members concatenated one after another. MultiGzDecoder decodes all members from the data and only returns Ok(0) when the underlying reader does. For a file, this reads to the end of the file.

To handle members seperately, see GzDecoder or read more in the introduction.

Examples

use std::io::prelude::*;
use std::io;
use flate2::bufread::MultiGzDecoder;

// Uncompresses a Gz Encoded vector of bytes and returns a string or error
// Here &[u8] implements BufRead

fn decode_reader(bytes: Vec<u8>) -> io::Result<String> {
   let mut gz = MultiGzDecoder::new(&bytes[..]);
   let mut s = String::new();
   gz.read_to_string(&mut s)?;
   Ok(s)
}

Implementations

impl<R: BufRead> MultiGzDecoder<R>

pub fn new(r: R) -> MultiGzDecoder<R>

Creates a new decoder from the given reader, immediately parsing the (first) gzip header. If the gzip stream contains multiple members all will be decoded.

impl<R> MultiGzDecoder<R>

pub fn header(&self) -> Option<&GzHeader>

Returns the current header associated with this stream, if itโs valid

pub fn get_ref(&self) -> &R

Acquires a reference to the underlying reader.

pub fn get_mut(&mut self) -> &mut R

Acquires a mutable reference to the underlying stream.

Note that mutation of the stream may result in surprising results if this decoder is continued to be used.

pub fn into_inner(self) -> R

Consumes this decoder, returning the underlying reader.