Module flate2::ffi::rust

Implementation for miniz_oxide rust backend.

Modules

deflate
This module contains functionality for compression.

inflate
This module contains functionality for decompression.

Structs

Deflate

Inflate

StreamResult
A structure containg the result of a call to the inflate or deflate streaming functions.

Enums

DataFormat
How compressed data is wrapped.

MZError
A list of miniz failed status codes.

MZFlush
A list of flush types.

MZStatus
A list of miniz successful status codes.

Constants

MZ_FINISH

MZ_FULL_FLUSH

MZ_NO_FLUSH

MZ_PARTIAL_FLUSH

MZ_SYNC_FLUSH

Functions

format_from_bool

Type Definitions

MZResult
Result alias for all miniz status codes both successful and failed.