Module flate2::gz

gzip compression/decompression

Structs

Builder
A builder structure to create a new gzip Encoder.

DecoderReader
A gzip streaming decoder

DecoderReaderBuf
A gzip streaming decoder

EncoderReader
A gzip streaming encoder

EncoderReaderBuf
A gzip streaming encoder

EncoderWriter
A gzip streaming encoder

Header
A structure representing the header of a gzip stream.

Statics

FCOMMENT

FEXTRA

FHCRC

FNAME

Functions

copy

corrupt