Module flate2::mem

Structs

Compress

Raw in-memory compression stream for blocks of data.

CompressError

Error returned when a compression object is used incorrectly or otherwise generates an error.

Decompress

Raw in-memory decompression stream for blocks of data.

DecompressError

Error returned when a decompression object finds that the input stream of bytes was not a valid input stream of bytes.

Enums

DecompressErrorInner 🔒

The inner state for an error when decompressing

FlushCompress

Values which indicate the form of flushing to be used when compressing in-memory data.

FlushDecompress

Values which indicate the form of flushing to be used when decompressing in-memory data.

Status

Possible status results of compressing some data or successfully decompressing a block of data.

Functions

compress_failed 🔒

decompress_failed 🔒

decompress_need_dict 🔒

write_to_spare_capacity_of_vec 🔒

Allows writer to write data into the spare capacity of the output vector. This will not reallocate the vector provided or attempt to grow it, so space for the output must be reserved by the caller before calling this function.