Module flate2::read

Types which operate over Read streams, both encoders and decoders for various formats.

Structs

DeflateDecoder

A DEFLATE decoder, or decompressor.

DeflateEncoder

A DEFLATE encoder, or compressor.

GzDecoder

A gzip streaming decoder

GzEncoder

A gzip streaming encoder

MultiGzDecoder

A gzip streaming decoder that decodes all members of a multistream

ZlibDecoder

A ZLIB decoder, or decompressor.

ZlibEncoder

A ZLIB encoder, or compressor.