Struct flate2::read::DeflateDecoder

pub struct DeflateDecoder<R: Read> {
    // some fields omitted
}

A DEFLATE decoder, or decompressor.

This structure implements a Read interface and takes a stream of compressed data as input, providing the decompressed data when read from.

Methods

impl<R: Read> DecoderReader<R>

fn new(r: R) -> DecoderReader<R>

Creates a new decoder which will decompress data read from the given stream.

fn new_with_buf(r: R, buf: Vec<u8>) -> DecoderReader<R>

Same as new, but the intermediate buffer for data is specified.

Note that the capacity of the intermediate buffer is never increased, and it is recommended for it to be large.

fn into_inner(self) -> R

Consumes this decoder, returning the underlying reader.