Struct flate2::read::DeflateDecoder

pub struct DeflateDecoder<R> { /* private fields */ }

A DEFLATE decoder, or decompressor.

This structure implements a Read interface and takes a stream of compressed data as input, providing the decompressed data when read from.

Examples

use std::io::prelude::*;
use std::io;
use flate2::read::DeflateDecoder;

// Uncompresses a Deflate Encoded vector of bytes and returns a string or error
// Here &[u8] implements Read
fn decode_reader(bytes: Vec<u8>) -> io::Result<String> {
   let mut deflater = DeflateDecoder::new(&bytes[..]);
   let mut s = String::new();
   deflater.read_to_string(&mut s)?;
   Ok(s)
}

Implementations

impl<R: Read> DeflateDecoder<R>

pub fn new(r: R) -> DeflateDecoder<R>

Creates a new decoder which will decompress data read from the given stream.

pub fn new_with_buf(r: R, buf: Vec<u8>) -> DeflateDecoder<R>

Same as new, but the intermediate buffer for data is specified.

Note that the capacity of the intermediate buffer is never increased, and it is recommended for it to be large.

impl<R> DeflateDecoder<R>

pub fn reset(&mut self, r: R) -> R

Resets the state of this decoder entirely, swapping out the input stream for another.

This will reset the internal state of this decoder and replace the input stream with the one provided, returning the previous input stream. Future data read from this decoder will be the decompressed version of r’s data.

Note that there may be currently buffered data when this function is called, and in that case the buffered data is discarded.

pub fn get_ref(&self) -> &R

Acquires a reference to the underlying stream

pub fn get_mut(&mut self) -> &mut R

Acquires a mutable reference to the underlying stream

Note that mutation of the stream may result in surprising results if this encoder is continued to be used.

pub fn into_inner(self) -> R

Consumes this decoder, returning the underlying reader.

Note that there may be buffered bytes which are not re-acquired as part of this transition. It’s recommended to only call this function after EOF has been reached.

pub fn total_in(&self) -> u64

Returns the number of bytes that the decompressor has consumed.

Note that this will likely be smaller than what the decompressor actually read from the underlying stream due to buffering.

pub fn total_out(&self) -> u64

Returns the number of bytes that the decompressor has produced.