Struct flate2::read::GzDecoder

pub struct GzDecoder<R: Read> { /* fields omitted */ }

A gzip streaming decoder

This structure exposes a Read interface that will consume compressed data from the underlying reader and emit uncompressed data.

Methods

impl<R: Read> DecoderReader<R>

fn new(r: R) -> Result<DecoderReader<R>>

Creates a new decoder from the given reader, immediately parsing the gzip header.

If an error is encountered when parsing the gzip header, an error is returned.

fn header(&self) -> &Header

Returns the header associated with this stream.