Struct flate2::Crc

pub struct Crc { /* fields omitted */ }

The CRC calculated by a CrcReader.

Methods

impl Crc

pub fn new() -> Crc

Create a new CRC.

pub fn sum(&self) -> u32

bla

pub fn amount(&self) -> u32

The number of bytes that have been used to calculate the CRC. This value is only accurate if the amount is lower than 232.

pub fn update(&mut self, data: &[u8])

Update the CRC with the bytes in data.

pub fn reset(&mut self)

Reset the CRC.

pub fn combine(&mut self, additional_crc: &Crc)

Combine the CRC with the CRC for the subsequent block of bytes.