Struct flate2::CrcReader

pub struct CrcReader<R> { /* fields omitted */ }

A wrapper around a Read that calculates the CRC.

Implementations

impl<R: Read> CrcReader<R>

pub fn new(r: R) -> CrcReader<R>

Create a new CrcReader.

impl<R> CrcReader<R>

pub fn crc(&self) -> &Crc

Get the Crc for this CrcReader.

pub fn into_inner(self) -> R

Get the reader that is wrapped by this CrcReader.

pub fn get_ref(&self) -> &R

Get the reader that is wrapped by this CrcReader by reference.

pub fn get_mut(&mut self) -> &mut R

Get a mutable reference to the reader that is wrapped by this CrcReader.

pub fn reset(&mut self)

Reset the Crc in this CrcReader.