Struct flate2::write::MultiGzDecoder

pub struct MultiGzDecoder<W: Write> { /* private fields */ }

A gzip streaming decoder that decodes all members of a multistream

A gzip member consists of a header, compressed data and a trailer. The gzip specification, however, allows multiple gzip members to be joined in a single stream. MultiGzDecoder will decode all consecutive members while GzDecoder will only decompress the first gzip member. The multistream format is commonly used in bioinformatics, for example when using the BGZF compressed data.

This structure exposes a [Write] interface that will consume all gzip members from the written buffers and write uncompressed data to the writer.

Implementations

impl<W: Write> MultiGzDecoder<W>

pub fn new(w: W) -> MultiGzDecoder<W>

Creates a new decoder which will write uncompressed data to the stream. If the gzip stream contains multiple members all will be decoded.

pub fn header(&self) -> Option<&GzHeader>

Returns the header associated with the current member.

pub fn get_ref(&self) -> &W

Acquires a reference to the underlying writer.

pub fn get_mut(&mut self) -> &mut W

Acquires a mutable reference to the underlying writer.

Note that mutating the output/input state of the stream may corrupt this object, so care must be taken when using this method.

pub fn try_finish(&mut self) -> Result<()>

Attempt to finish this output stream, writing out final chunks of data.

Note that this function can only be used once data has finished being written to the output stream. After this function is called then further calls to write may result in a panic.

Panics

Attempts to write data to this stream may result in a panic after this function is called.

Errors

This function will perform I/O to finish the stream, returning any errors which happen.

pub fn finish(self) -> Result<W>

Consumes this decoder, flushing the output stream.

This will flush the underlying data stream and then return the contained writer if the flush succeeded.

Note that this function may not be suitable to call in a situation where the underlying stream is an asynchronous I/O stream. To finish a stream the try_finish (or shutdown) method should be used instead. To re-acquire ownership of a stream it is safe to call this method after try_finish or shutdown has returned Ok.

Errors

This function will perform I/O to complete this stream, and any I/O errors which occur will be returned from this function.