Struct flate2::write::ZlibDecoder

pub struct ZlibDecoder<W: Write> { /* private fields */ }

A ZLIB decoder, or decompressor.

This structure implements a Write and will emit a stream of decompressed data when fed a stream of compressed data.

Examples

use std::io::prelude::*;
use std::io;
use flate2::write::ZlibDecoder;

// Uncompresses a Zlib Encoded vector of bytes and returns a string or error
// Here Vec<u8> implements Write

fn decode_reader(bytes: Vec<u8>) -> io::Result<String> {
   let mut writer = Vec::new();
   let mut z = ZlibDecoder::new(writer);
   z.write_all(&bytes[..])?;
   writer = z.finish()?;
   let return_string = String::from_utf8(writer).expect("String parsing error");
   Ok(return_string)
}

Implementations

impl<W: Write> ZlibDecoder<W>

pub fn new(w: W) -> ZlibDecoder<W>

Creates a new decoder which will write uncompressed data to the stream.

When this decoder is dropped or unwrapped the final pieces of data will be flushed.

pub fn new_with_decompress(w: W, decompression: Decompress) -> ZlibDecoder<W>

Creates a new decoder which will write uncompressed data to the stream w using the given decompression settings.

When this decoder is dropped or unwrapped the final pieces of data will be flushed.

pub fn get_ref(&self) -> &W

Acquires a reference to the underlying writer.

pub fn get_mut(&mut self) -> &mut W

Acquires a mutable reference to the underlying writer.

Note that mutating the output/input state of the stream may corrupt this object, so care must be taken when using this method.

pub fn reset(&mut self, w: W) -> Result<W>

Resets the state of this decoder entirely, swapping out the output stream for another.

This will reset the internal state of this decoder and replace the output stream with the one provided, returning the previous output stream. Future data written to this decoder will be decompressed into the output stream w.

Errors

This function will perform I/O to complete this stream, and any I/O errors which occur will be returned from this function.

pub fn try_finish(&mut self) -> Result<()>

Attempt to finish this output stream, writing out final chunks of data.

Note that this function can only be used once data has finished being written to the output stream. After this function is called then further calls to write may result in a panic.

Panics

Attempts to write data to this stream may result in a panic after this function is called.

Errors

This function will perform I/O to complete this stream, and any I/O errors which occur will be returned from this function.

pub fn finish(self) -> Result<W>

Consumes this encoder, flushing the output stream.

This will flush the underlying data stream and then return the contained writer if the flush succeeded.

Note that this function may not be suitable to call in a situation where the underlying stream is an asynchronous I/O stream. To finish a stream the try_finish (or shutdown) method should be used instead. To re-acquire ownership of a stream it is safe to call this method after try_finish or shutdown has returned Ok.

Errors

This function will perform I/O to complete this stream, and any I/O errors which occur will be returned from this function.

pub fn total_in(&self) -> u64

Returns the number of bytes that the decompressor has consumed for decompression.

Note that this will likely be smaller than the number of bytes successfully written to this stream due to internal buffering.

pub fn total_out(&self) -> u64

Returns the number of bytes that the decompressor has written to its output stream.